Adding a way to externally refresh the pages
I am loading data externally which populates correctly in the table but when I first load the data via ajax and set the total on the ngTableParams the pages collection is not updated.  It would be nice if it were updated automatically but at the least a method to trigger the reload should be added.