Fix(header.html): sort icon appears twice using custom header template
The template for header cell causes the sort icon to appear twice when a custom header template is requested (eg data-header='myColumnHeader.html').

The header template introduces a redudant nested div in the case of a custom header template. This extra div is then targed by a css descendant selector that adds the sort icon.

This fix removes the redundant div thus avoiding the css issue.