feat(header): header-class attribute is now a data binding expression
Previously the header-class attribute was added to a th element as a string literal.
Now the header-class attribute is supplied as a binding expression which has access
to the column definition.

This is a start in a move toward a design where TD element attributes are expressions
that have access to the scope.

BREAKING CHANGE:

Previously, a css class was added to TH elements thusly:

````
<tr ng-repeat="row in $data">
	<td header-class="myHeaderClass"></td>
</tr>
````

Now:

````
<tr ng-repeat="row in $data">
	<td header-class="'myHeaderClass'"></td>
</tr>
````