fix(ngTableController): prevent "stackoverflow" exception when data items are recursive data structures
Watching the entire ngTableParams structure "by value" including it's assigned page of data causes
angular.copy to throw an exception.

This fix reduces that part of the ngTableParams structure that is observed for changes.

The fix should also improve performance where bound objects are large graphs and/or the grid
page size is large