Pager: Hiding irrelevant page size controls
Page size controls are always displayed even if they are irrelevant to current dataset.  I..e Say we have 10, 20, 50 page size controls.  There is no need to show any when there is only 3 items on the list.  There is also no need to show 50 if there are only 15 items in the list.