fix(ngTableController): don't trigger reload whilst a reload is already in-flight
Disabling the `watch` of changes to the table `filter` (etc) whilst a reload is
inprogress allows us to correctly *chain* multiple calls to reload.

It allows the caller to wait on a promise that will be resolved when all calls to reload
are completed.

BREAKING CHANGE:

Calls to `NgTableParams.filter`, `NgTableParams.sorting` (etc) made in the `then` method of
the promise returned by `NgTableParams.reload` will NOT trigger a subsequent call to `NgTableParams.reload`;
the call to `NgTableParams.reload` must now be explicitly be made.

Previously:

````js
tableParams.reload().then(function(){
  if (!tableParams.total() && _.size(tableParams.filter()) > 0) {
	tableParams.filter({});
  }
});
````

Now:

````js
tableParams.reload().then(function(){
  if (!tableParams.total() && _.size(tableParams.filter()) > 0) {
	tableParams.filter({});
	return tableParams.reload();
  }
});
````