feat(ngTableFilterConfig): allow template urls for filters to be customized
`ngTableFilterConfig` is a new service that will allow various aspects of ng-table filters to be
customised.

Out of the box, the service allows an application to configure which url should be selected as the
template url for a filter.

**Example: override the url selected for the 'text' filter:**

```js
angular.module('yourApp').config(function(ngTableFilterConfigProvider){
    ngTableFilterConfigProvider.setConfig({ aliasUrls: {
        'text': 'custom/url/custom-text.html'
    }});
});
```

**Example: override the *default* url for all filter templates**

```js
angular.module('yourApp').config(function(ngTableFilterConfigProvider){
    ngTableFilterConfigProvider.setConfig({ defaultBaseUrl: 'default/path/to/filter/templates/'});
});
```