docs(demos): remove unsupported pagination example
I'm not sure how the original example of external pagination worked - it looks to me that
you must always provide a templateUrl to the pagination directive.

Also external pagination doesn't really work - it needs an option to configure the `NgTableParams`
to allow for the standard pager for the table to be switched off.