fix(ngTableController): should be consistent about adding empty option item
When the result returned by `filterData` is a promise that resolves to an array, an empty option
is added, but not when `filterData` returns an array synchronously.

This commit ensures that an empty option is added regardless of whether an array is returned
synchronously or asynchronously.

BREAKING CHANGE:

An empty item - `{ id: '', title: '' }` - is added to an array returned synchronously by `filterData`
function.

Implications:
* make sure to *not* add an empty option yourself as this will be a duplicate
* your array of items need to have an `id` and `title` field so as to match the empty option