fix(NgTableParams): filterDelay too high
Trying to get a reasonable balance between responsiveness and not wasting calls to a server.

A better approach would be for a filter template to be able to specify a filter delay and this be used
in an ngModelOptions. That would allow a template like a select to not delay at all whilst a text
input to add a delay for typical typing speeds.