fix(filterRow.html): header-class should also apply to filter row
BREAKING CHANGE:

A css class specified using the header-class will now be added to the filter row header cell and not
just the sorting row.

If you want to continue to apply the css rules *only* to the cell in the sorting header row you
will now need to qualify your css rule with the '.header' css class.

So the following:

```css
.my-customer-header {
    /* rules */
}
```

... will need to change to:

```css
.header.my-customer-header {
    /* rules */
}
```