chore(npm):  stop implicitly installing typings
Using postinstall or prepublish npm script to install typings
does not work in practice: it will cause this package to fail to
install as a depedency of a consuming package when the consumer
does not have typings installed.

What we really want is a npm lifecycle hook that will run a script
only when running npm install on this cloned github repo ie NOT
when this package is installed as a dependency of another package.

That would support a very simple install a dev setup:
1. clone this repo
2. `npm install`

Until such times there is an extra step required:
1. clone this repo
2. `npm install`
3. `npm run init`