feat(ng-table.js): support the CommonJS module format
ng-table.js is now output in UMD format and thus supports the following module formats:
* global script tag
* AMD (eg RequireJS)
* CommonJS

BREAKING CHANGE:

1) The AMD (RequireJS) module returned by ng-table.js is now an object that references
the angular module.

Previously, the module returned *was* the angular module.

Note: this will only affect apps that are using RequireJS to load ng-table.js.

Those apps that are loading ng-table.js as a script tag will be unaffected.

2) Replaced LESS with SASS

This will only affect the minority of apps that are using the ng-table.less file in the
dist folder.

Those apps that are using the ng-table.css will be unaffected.