Text filter doesn't autocomplete
I don't know if this is the desired performance for others but for me I don't want autocomplete results to show up in the text filter. I have only tested in Firefox but when starting to implement this into my project I noticed that all of my autocomplete results would load as I started typing which obstructed the data below that was getting filtered and I didn't want that.